fix: Fixes window undefined in jsx/react webworkers
Previously the HMR code would use window as the global scope which is undefined in webworkers.

This change uses some code lifted from @vitejs/plugin-react to check if the context is in a webworker before doing trying to use `window.

HMR is not currently possible in vite (https://github.com/vitejs/vite-plugin-react/issues/152) so nothing happens when the worker file is edited.